#173 fixed merged_to_duty error
